requirements.txt
Flask
python-dateutil
requests
google-cloud-secret-manager
